% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotreg.r
\name{plotreg}
\alias{plotreg}
\title{Returns scatter plot with regression results}
\usage{
plotreg(x, y, dat, titl, xlab, ylab, scx, scy, w, names, sepa, ini, font,
  logs, cv, nx, ny, stepx, stepy)
}
\arguments{
\item{x}{string with name of variable on x axis}

\item{y}{string with name of variable on y axis}

\item{dat}{data frame with data}

\item{titl}{string with tittle}

\item{xlab}{string with lable x axis, x by default}

\item{ylab}{string with lable y axisa, y by default}

\item{scx}{scale x variable. E.g 10s,100s, 1000s,... scx=1 by default}

\item{scy}{scale y variable. E.g 10s,100s, 1000s,... scy=1 by default}

\item{w}{string with name of regression weights variable. Equal weight by default}

\item{names}{string with name of data points labels. E.g. City, State, Country}

\item{sepa}{parameter used to calibrate the frequency tick marks,  large values imply lower frequency. sepa=0 produces one label per data point}

\item{ini}{number beyond which the frequency of names is reduced}

\item{font}{plot font, serif by default}

\item{logs}{use of logs in the regression, logs=0 y~x, logs=1: log(y)~log(x), logs=2: log(y)~x, logs=3: y~log(x)}

\item{cv}{name of covariates}

\item{nx}{horizontal adjustment for data points labels}

\item{ny}{vertical adjustment for data points labels}
}
\value{
weighted scater plot with weighted linear regression
}
\description{
Returns scatter plot with regression results
}
\examples{
 
exmp<-plotreg(x='Distance_from_Largest_City_to_NYC_miles',y='Total_Trade_Amount_2015_Thousand_USD',dat=trade,titl='trade',xlab = 'distance',ylab='trade',w='GDP_2015_USD',names='Name_country',nx=100,ny=0)
exmp
}
