% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtable.r
\name{regtable}
\alias{regtable}
\title{Returns a matrix and LaTex table with results from a set of linear regressions (using lm())}
\usage{
regtable(form, datav, namescol, namesrow, stars, wgh, rbst, clst, dec)
}
\arguments{
\item{form}{list with formulas}

\item{datav}{data for regressions}

\item{namescol}{tittles of regression}

\item{namesrow}{regressors name}

\item{stars}{TRUE for stars in the table}

\item{wgh}{name of variable used to weight regression}

\item{rbst}{TRUE for robust errors}

\item{clst}{name of variable used to cluster errors}

\item{dec}{number of decimals in the table}
}
\value{
$matrix returns a matrix with all the results from the regression, $latex returns a latex table with all the results from the regression
}
\description{
Returns a matrix and LaTex table with results from a set of linear regressions (using lm())
}
\examples{
form<-list()
form[[1]]<-log(Total_Trade_Amount_2015_Thousand_USD)~log(GDP_2015_USD)
form[[2]]<-log(Total_Trade_Amount_2015_Thousand_USD)~log(GDP_2015_USD)+log(Distance_from_Largest_City_to_NYC_miles)
namescol<-c('GDP 2015', 'GDP & Distance')
namesrow<-c('GDP 2015', 'Distance')
lmtable_res<-regtable(form,datav=trade,namescol,namesrow)
mtrx<-lmtable_res$matrix
lmtable_res$latex
}
