% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipfslaw.r
\name{zipfslaw}
\alias{zipfslaw}
\title{Returns plot of Zip's law regression}
\usage{
zipfslaw(data, var, name, namevar, sepa, color, width, pointsize, height,
  path)
}
\arguments{
\item{data}{data set}

\item{var}{variable name string}

\item{name}{lable observation}

\item{namevar}{name string to equation}

\item{sepa}{higher numbers reduce the frequency at which observation names apper}

\item{color}{text color}

\item{width}{figure width}

\item{pointsize}{size data points}

\item{height}{height plot}

\item{path}{filepath where the output is stored (e.g. path='C:/Users/jj/R'). If missing is stored at the current loaction (getwd()). filed is saved as Zips_+namevar}
}
\value{
png file stored in path
}
\description{
Returns plot of Zip's law regression
}
\examples{
 
zipfslaw(data=trade,var='GDP_2015_USD',name='Name_country',namevar = 'GDP',sepa=3,path='C:/Users/jj/file')
}
